backend/requirements.txt
awesome-slugify==1.6.5
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.14.6
charset-normalizer==2.0.7
click==8.0.1
colorama==0.4.4
contourpy==1.2.0
cycler==0.12.1
dnspython==2.1.0
email-validator==1.1.3
Faker==20.1.0
Flask==2.0.1
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
fonttools==4.46.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.2
importlib-resources==6.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
kiwisolver==1.4.5
lorem==0.1.1
MarkupSafe==2.0.1
matplotlib==3.8.2
networkx==3.2.1
numpy==1.26.2
packaging==23.2
Pillow==8.4.0
pycparser==2.20
PyJWT==2.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==0.19.0
regex==2021.11.10
requests==2.26.0
scikit-fuzzy==0.4.2
scipy==1.11.4
six==1.16.0
SQLAlchemy==1.4.25
Unidecode==0.4.21
urllib3==1.26.7
Werkzeug==2.0.1
WTForms==2.3.3
zipp==3.17.0
